+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.2.0] - X
 ### Added
+- Ocean Watch: Ocean Watch beta. [OW-171](https://vizzuality.atlassian.net/browse/OW-171)
 - Ocean Watch: analytics events. [OW-76](https://vizzuality.atlassian.net/browse/OW-76)
 - Ocean Watch: added OW partners carousel to several pages.
 - Ocean Watch: populates about page with API content. [OW-63](https://vizzuality.atlassian.net/browse/OW-63)